requirements.txt
requests==2.22.0
python-dateutil==2.8.0
pytz==2019.2
cfscrape==2.1.1

